Removing variate campaign testing
In recent history, MailChimp implemented a new policy (https://stackoverflow.com/questions/56426392/mailchimp-api-3-0-status-403-user-does-not-have-access-to-the-requested-operat) that makes it impossible for a free user like myself to test the multivariate campaign feature. It should still be supported (with automatic field-verification), but that verification cannot be checked with Jest in this repo.